requirements.txt
streamlit==1.27.2
Pillow==8.4.0  # or any other packages your app requires
